fix(ci): split YouTube seeking helpers
